Final iteratation
Added Ui elements to explain how to do the feature mechanics, removed visible traces and print statements that were used for debugging, added some more comments around much of the code to explain it better and tidy it up. Also added the charges mechanic and the refill mechanic and have this tracked on the UI. The player can pause the game and quit from there.